add option to keep all media fragment files
Sometimes it is useful to be able to test an encoded stream by just
hosting the manifest and media fragments on an HTTP server. For that
usecase, keeping each init segment and media segment in the "dash"
directory is needed. This commit adds a "--preserve" option that will
cause the packaging tasks to keep all these intermediate files.

+++ dashlive/media/create/task.py
@@ -139,7 +139,8 @@ class MediaCreationTask(ABC):
                     sys.stdout.flush()
                 with open(moof, "rb") as src:
                     shutil.copyfileobj(src, dest)
-                os.remove(moof)
+                if not self.options.preserve:
+                    os.remove(moof)
                 segment += 1
             if self.options.verbose:
                 sys.stdout.write('\n')
